escrow-backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
gunicorn==20.1.0
amqp==5.1.1
anyio==3.7.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
cryptography==41.0.2
Django==4.2.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
django-timezone-field==5.1
djangorestframework==3.14.0
dnspython==2.4.0
eventlet==0.33.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
idna==3.4
kombu==5.3.1
oauthlib==3.2.2
prompt-toolkit==3.0.39
pycparser==2.21
PyJWT==2.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
stripe==5.4.0
tweepy==4.14.0
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
